add / fixe %u %o %x %p

+++ libft/ft_nblen.c
@@ -6,18 +6,18 @@
 /*   By: rpinoit <rpinoit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/21 11:05:38 by rpinoit           #+#    #+#             */
-/*   Updated: 2018/03/21 14:12:45 by rpinoit          ###   ########.fr       */
+/*   Updated: 2018/03/23 10:45:27 by rpinoit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_nblen(intmax_t n, intmax_t base)
+int		ft_nblen(intmax_t n, short base)
 {
 	int len;
 
 	len = 1;
-	while (n /= base)
+	while (n /= (intmax_t)base)
 		++len;
 	return (len);
 }
